+++ README.md
@@ -1,23 +1,36 @@
-# HCI
-Histórico Clínico Integrado
+This is a [Next.js](https://nextjs.org/) project bootstrapped with [`create-next-app`](https://github.com/vercel/next.js/tree/canary/packages/create-next-app).
 
-## Primeiros Passos
+## Getting Started
 
-1. #### Instale as dependências
-Na raiz do projeto, rode o comando abaixo:
+First, run the development server:
 
 ```bash
-pnpm i
+npm run dev
+# or
+yarn dev
+# or
+pnpm dev
+# or
+bun dev
 ```
 
-2. #### Forneça as variáveis ambiente
-Crie o arquivo `.env.local` e forneça as variáveis ambientes necessárias. Utilize o arquivo `.env.example` para saber quais variáveis são necessárias.
+Open [http://localhost:3000](http://localhost:3000) with your browser to see the result.
 
-3. #### Inicie o servidor de desenvolvimento
-Na raiz do projeto, rode o comando abaixo:
+You can start editing the page by modifying `app/page.tsx`. The page auto-updates as you edit the file.
 
-```bash
-pnpm run dev
-```
+This project uses [`next/font`](https://nextjs.org/docs/basic-features/font-optimization) to automatically optimize and load Inter, a custom Google Font.
 
-Abra [http://localhost:3000](http://localhost:3000) no seu navegador para ver o resultado.
+## Learn More
+
+To learn more about Next.js, take a look at the following resources:
+
+- [Next.js Documentation](https://nextjs.org/docs) - learn about Next.js features and API.
+- [Learn Next.js](https://nextjs.org/learn) - an interactive Next.js tutorial.
+
+You can check out [the Next.js GitHub repository](https://github.com/vercel/next.js/) - your feedback and contributions are welcome!
+
+## Deploy on Vercel
+
+The easiest way to deploy your Next.js app is to use the [Vercel Platform](https://vercel.com/new?utm_medium=default-template&filter=next.js&utm_source=create-next-app&utm_campaign=create-next-app-readme) from the creators of Next.js.
+
+Check out our [Next.js deployment documentation](https://nextjs.org/docs/deployment) for more details.
