+++ src/app/(app)/[cpf]/componentes/patient-details/components/basic-info/medications-and-alergies/components/medications-and-alergies-popover.tsx
