+++ src/app/(app)/components/cpf-search/index.tsx
@@ -2,7 +2,7 @@ import { CPFSearchForm } from './components/cpf-search-form'
 
 export function CPFSearch() {
   return (
-    <main className="mx-auto flex w-full max-w-[36.4375rem] translate-y-10 flex-col items-center justify-center p-4">
+    <main className="mx-auto flex h-screen w-full max-w-[36.4375rem] flex-col items-center justify-center p-4">
       <div className="flex flex-col items-center gap-3 p-6">
         <h2 className="text-center text-2xl font-semibold leading-6 tracking-tight text-primary">
           Buscar por CPF
