+++ src/app/(app)/page.tsx
@@ -1,11 +1,9 @@
 import { CPFSearch } from './components/cpf-search'
-import { UserNotice } from './components/cpf-search/components/user-notice'
 
 export default function Home() {
   return (
-    <div className="relative flex h-screen w-full min-w-screen-sm flex-col items-center justify-center gap-10">
+    <div className="w-full min-w-screen-sm">
       <CPFSearch />
-      <UserNotice />
     </div>
   )
 }
