+++ tailwind.config.ts
@@ -35,7 +35,6 @@ const config = {
           200: '#F6F6F6',
           300: '#F3F3F3',
         },
-        'dark-gray': '#3d3f40',
         'light-green': '#DBEED2',
         'light-yellow': '#EfE3B5',
         'sky-blue': '#CCE0EA',
